Set common deployment targets for the whole Xcode project
Setting the iOS deployment target at the project level ensures the IPHONEOS_DEPLOYMENT_TARGET environment variable is available to the scripts which build the external libraries.
Before this change, linking the external libraries with SwiftGit2-iOS resulted in hundreds of warnings that "object file (...) was built for newer iOS version (10.1) than being linked (9.2)".