Clean up static library build settings (and the names of the products)
The static library dependencies are built by external scripts, and do not use the naming settings specified here. However, Xcode uses the PRODUCT_NAME setting to generate the name of the placeholder product reference it creates automatically for each target.Specifying the base names of each static library in the build settings causes Xcode to generate names for its placeholder products which match the products that are actually built by the scripts.