Fix build on macOS 15 with Xcode 16
A compilation error was caused by the zlib dependency that is bundled
with libgit2. This particular version of zlib fails to compile on
Clang 17 with the error "Expected identifier or '('".
This problem was fixed in libgit2 v1.9.0 in which they updated zlib.

However, libgit2 v1.9.0 introduced a new compilation issue:
an unterminated multiline comment, which is fixed in this
commit which is not part of any libgit2 release yet:
https://github.com/libgit2/libgit2/commit/2fa13adf09c8dd1f334f57023390043ea3f71cb2

When I check out the libgit2 submodule at that commit
(and remove a reference to a deprecated constant from SwiftGit2) it is solved.

Changes in this commit:

* Updates libgit2 to 2fa13adf09c8dd1f334f57023390043ea3f71cb2, this is
  needed to fix the compilation issues.
* Removes reference to deprecated constant `GIT_OBJECT_REF_DELTA`, which
  wasn't being used in SwiftGit2.